Feature/optimize sql queries

+++ update-boundary-admin.sh
@@ -25,24 +25,29 @@ function displaytime {
 # Load config
 if [ -f config.cfg ] ; then
 	source config.cfg
+	echo -e "${Gra}Config : ${Gre}OK${RCol}"
 else
 	echo -e "\e[1;31mPlease configure the script by renaming the file 'config.defaut.cfg' to 'config.cfg.\e[0m"
 	exit;
 fi
 
 # Check if the osm-c tools are needed
-echo -e "${Yel}Downloading and building the osm-c tools if necessary...${RCol}";
-cd $DIR_BIN
-if [ ! -f ${DIR_BIN}/osmupdate ] ; then
-	wget -O - http://m.m.i24.cc/osmupdate.c | cc -x c - -o osmupdate
+if [ ! -f ${DIR_BIN}/osmupdate ] || [ ! -f ${DIR_BIN}/osmconvert ] || [ ! -f ${DIR_BIN}/osmfilter ] ; then
+	echo -e "${Yel}Downloading and building the osm-c tools...${RCol}";
+	cd $DIR_BIN
+	if [ ! -f ${DIR_BIN}/osmupdate ] ; then
+		wget -O - http://m.m.i24.cc/osmupdate.c | cc -x c - -o osmupdate
+	fi
+	if [ ! -f ${DIR_BIN}/osmconvert ] ; then
+		wget -O - http://m.m.i24.cc/osmconvert.c | cc -x c - -lz -O3 -o osmconvert
+	fi
+	if [ ! -f ${DIR_BIN}/osmfilter ] ; then
+		wget -O - http://m.m.i24.cc/osmfilter.c |cc -x c - -O3 -o osmfilter
+	fi
+	cd $DIR_BASE
+else
+	echo -e "${Gra}OSM C tools : ${Gre}OK${RCol}"
 fi
-if [ ! -f ${DIR_BIN}/osmconvert ] ; then
-	wget -O - http://m.m.i24.cc/osmconvert.c | cc -x c - -lz -O3 -o osmconvert
-fi
-if [ ! -f ${DIR_BIN}/osmfilter ] ; then
-	wget -O - http://m.m.i24.cc/osmfilter.c |cc -x c - -O3 -o osmfilter
-fi
-cd $DIR_BASE
 
 # Check if the ogr2ogr was built with support of Mysql and OSM
 if [ ! -f ${DIR_BIN_OGR}/ogr2ogr ] &&	 [ ! -f ${DIR_BIN_OGR}/ogrinfo ] ; then
@@ -53,7 +58,7 @@ else
 		echo -e "${Red}Your version of GDAL does not support OSM and MySQL formats.${RCol}"
 		exit;
 	else
-		echo -e "${Gre}GDAL supports Mysql and OSM.${RCol}"
+		echo -e "${Gra}GDAL supports Mysql and OSM : ${Gre}OK${RCol}"
 	fi
 fi
 
@@ -64,7 +69,7 @@ do
 
 	# Maintain up to date the .osm.pbf file
 	if [ ! -f "${DIR_OSM}/${AREA}.osm.pbf" ] ; then
-		echo -e "${Yel}Téléchargement du fichier PBF initial pour la zone «${AREA}»...${RCol}";
+		echo -e "${Yel}Downloading initial PBF file for area «${AREA}»...${RCol}";
 		if [ $AREA == "france" ] ; then
 			URL="http://download.geofabrik.de/europe/${AREA}-latest.osm.pbf"
 		else
@@ -74,12 +79,12 @@ do
 	else
 		# Check if an update has been made less than 8 hours
 		if [ `ageEnSeconde "${DIR_OSM}/${AREA}.osm.pbf"` -gt 28800 ] ; then
-			echo -e "${Yel}Mise à jour du fichier PBF pour la zone «${AREA}»...${RCol}";
+			echo -e "${Yel}Updating the PBF file for area «${AREA}»...${RCol}";
 			mv ${DIR_OSM}/${AREA}.osm.pbf ${DIR_OSM}/${AREA}_old.osm.pbf
 			${DIR_BIN}/osmupdate -B=${DIR_POLY}/${AREA}.poly -v -t=${DIR_TMP}/osmupdate/temp --day ${DIR_OSM}/${AREA}_old.osm.pbf ${DIR_OSM}/${AREA}.osm.pbf
 			rm -f ${DIR_OSM}/${AREA}_old.osm.pbf
 		else
-			echo -e "${Gre}${AREA}.osm.pbf up to date${RCol}";
+			echo -e "${Gre}${AREA}.osm.pbf is up to date${RCol}";
 		fi
 	fi
 
@@ -89,7 +94,7 @@ do
 		exit 1;
 	else
 		if [ ! -f "${DIR_OSM}/${AREA}_boundary.osm.pbf" ] || [ `ageEnSeconde "${DIR_OSM}/${AREA}_boundary.osm.pbf"` -gt 28800 ] ; then
-			echo -e "${Yel}Filtrage des zones administratives en cours pour la zone «${AREA}»...${RCol}"
+			echo -e "${Yel}Filtering the administrative boundaries for area «${AREA}»...${RCol}"
 			${DIR_BIN}/osmconvert ${DIR_OSM}/${AREA}.osm.pbf --out-o5m > ${DIR_OSM}/${AREA}.o5m
 			${DIR_BIN}/osmfilter -t=${DIR_TMP}/osmfilter_tempfile \
 				${DIR_OSM}/${AREA}.o5m \
@@ -100,7 +105,7 @@ do
 				--out-o5m | ${DIR_BIN}/osmconvert - --out-pbf -o=${DIR_OSM}/${AREA}_boundary.osm.pbf
 			rm -f ${DIR_OSM}/${AREA}.o5m
 		else
-			echo -e "${Gre}${AREA}_boundary.osm.pbf up to date${RCol}";
+			echo -e "${Gre}${AREA}_boundary.osm.pbf is up to date${RCol}";
 		fi
 	fi
 
@@ -109,7 +114,7 @@ do
 		echo -e "${Red}Can not find the file : ${AREA}_boundary.osm.pbf${RCol}";
 		exit 1;
 	else
-		echo -e "${Yel}Importing into MySQL for the area «${AREA}»...${RCol}";
+		echo -e "${Yel}Importing into MySQL for area «${AREA}»...${RCol}";
 		${DIR_BIN_OGR}/ogr2ogr \
 			--config OSM_CONFIG_FILE $OSM_CONF_INI \
 			--config MYSQL_UNIX_PORT $MYSQL_UNIX_PORT \
@@ -122,7 +127,7 @@ do
 	fi
 
 	# Create and maintain the ref table
-	echo -e "${Yel}Update the reference table for the zone «${AREA}»...${RCol}"
+	echo -e "${Yel}Updating the reference table for area «${AREA}»...${RCol}"
 	$BIN_PHP ./update-boundary-admin-ref.php ${AREA}
 
 	# Show time elapsed
